requirements.txt
streamlit
librosa
tensorflow-cpu
numpy
pandas
scipy
scikit-learn
matplotlib